python/examples/mcp/requirements.txt
lightspark
aiohttp
mcp
python-dotenv
